UpdatesToGameRatingUI
Updates to game rating UI

+++ app/Services/UserService.php
@@ -17,7 +17,7 @@ class UserService
         return $this->userRepository->getWishlistByUserId();
     }
 
-    public function submitGameReview(int $game_id, int $user_id, int $game_rating, string $text_review):Reviews
+    public function submitGameReview(int $game_id, int $user_id, float $game_rating, string $text_review):Reviews
     {
         $submittedReview = $this->userRepository->InsertGameReview($game_id, $user_id, $game_rating, $text_review);
         $returnedReview =  new Reviews();
